メールグループのstoreを作成
